Support longer path-based services, make domain default
The use case for domain routing is really as a default to go to in case
no service is specified in the path. Further, a path->service may
require a longer path than the first component of a path, e.g.
/api/service vs /service. This also applies to subdomains->service, e.g.
sub.domain.com vs sub.sub2.domain.com.

This adds longest-substring matching for subdomain and path matches so
that we specify an entire subpath or entire subdomain to match proper
service name lookup.

Example:
  service_mappings = {["/api"]="service_api", ["/api/service1"]="service1"}
  subdomain_mappings = {["api.subdomain"]="service1", api="service_api"}

Maps the following URIs to their service name:
  api.example.com/ => service_api
  api.example.com/api/service1 => service1
  api.subdomain.example.com/ => service1
  api.subdomain.example.com/api => service_api