Add external aggregate health checks in JSON
Border Patrol currently depends on several upstream services for it to
work properly. Since the lack of certain services renders it useless, we
should report that in our health checks.

Further, the usefulness of CRUD checks in memcached is proving to be
more trouble than it's worth. Instead, it will report the stats for
memcached