When redirecting user with session, if no redirect url redirect to default
For users without a borderpatrol session we store in memcache the url
they were going to so that we can redirect them there after logging in. If
a logged in user accesses /a, since we deleted the redirect url when they
got the session, we end up redirecting to "". If we are about to redirect
to "" redirect to the default for the service. If there is no default
redirect to "/".

It also seems that different browsers handle redirects to "" differently.
Chrome does nothing, while Safari and Firefox redirect to "/"